==> doc/01-processo.Rmd <==
---
title: "Processo"
subtitle: "res431"
author: "vincenzo.matarazzo@cogne.com"
version: 1.0 
date: "`r Sys.Date()`"
output:
  rmdformats::readthedown:
    css: ./custom.css
    df_print: paged
    gallery: no
    highlight: default
    html_document: null
    lightbox: yes
    number_sections: yes
    self_contained: yes
    thumbnails: no
editor_options:
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      warning = FALSE,
                      message = FALSE,
                      error = FALSE,
                      fig.height = 7,
                      fig.width = 10,
                      collapse = TRUE,
                      cols.print=20)
rm(list = ls(all.names = TRUE))
```



# Il problema

I risultati di *resilienza* che stiamo ottenendo sul *431/5* sono instabili e spesso fuori specifica (sotto il limite **minimo di 50 J**). In caso di schede fuori specifica, la scheda viene ritrattata. 

## workflow processo

ACC produce billette 160x160. Le billette entrano successivamente al forno BENDOTTI (storicamente, set-point a 1100°C per garantire le caratteristiche meccaniche). A seguire, sbozzatura da quadro 160 a sbozzato 87. 

Gli sbozzati entrano quindi al treno LAM, dove vengono laminati a barre di diametro 47 dopodiché vengono raffreddati in placca (aria/acqua), tagliate a lunghezza circa costante e confezionate/stoccate in pallets prima di affrontare una campagna di trattamenti termici.

I trattamenti termici previsti (TT) sono 2 cicli di rinvenimento (GADDA, forni a batch). La capacità di questi forni è di circa 15 tonnellate. Per quanto riguarda i TT, **si fa l'ipotesi che i cicli siano privi di anomalie, e sempre in specifica**.  

Il trattamento termico prevede 2 cicli di rinvenimento, così costituiti:

+ rinvenimento 1 (monostrato)

  + $T_{start}$ : 300 °C
  + rampa 100 °C/h fino a 500 °C
  + rampa 50 °C/h fino a 660 °C
  + mantenimento x 6h

+ aria ventilata fino a $T_{amb}$ (circa 4-6 h)

+ preparazione dei fasci di barre per il rinvenimento 2 (fascettatura)

+ rinvenimento 2 (in fasci)

  + $T_{start}$ : 300 °C
  + rampa 100 °C/h fino a 500 °C
  + rampa 50 °C/h fino a 580 °C
  + mantenimento x 6h

+ raffreddamento in vasca (H20) fino a $T_{amb}$, con un minimo di 20 min. 


Il campionamento per la prova a trazione (Kv) è effettuato su una barra da cui vengono ricavati 3 provini. I valori di resilienza quindi sono 3 per ogni sk. 

**Le schede ritrattate NON sono state considerate**
