==> doc/03-data-mapping.Rmd <==
---
title: "Data Mapping"
subtitle: "res431"
author: "philippe.brunier@cogne.com"
version: 1.0 
date: "`r Sys.Date()`"
output:
  rmdformats::readthedown:
    css: ./custom.css
    df_print: paged
    gallery: no
    highlight: default
    html_document: null
    lightbox: yes
    number_sections: yes
    self_contained: yes
    thumbnails: no
editor_options:
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      warning = FALSE,
                      message = FALSE,
                      error = FALSE,
                      fig.height = 7,
                      fig.width = 10,
                      collapse = TRUE,
                      cols.print=20)
rm(list = ls(all.names = TRUE))

require(dplyr)
require(kableExtra)
require(readr)
require(reticulate)
source('~/dev/res431/R/function.R')
```

# File

I file utilizati in questo progetto sono: 

1. dati_CH.txt
1. dati_KV_TVB.txt
1. dati_G14.txt
1. datiRM.txt
1. dati_RINV1.txt
1. dati_RINV2v2.txt

## dati_CH 

Il file i finali di colata AOD, con le relative chimiche prese da APC (rcchim02)

```{r head_ch, include = TRUE}

header <- read_delim('/data/res431/header/header_CH.csv', delim=';', col_type=cols())

header <- filter(header, grab>0)
header <- header[-ncol(header)]

```

Il file è composto da `r nrow(header)` colonne, con le seguenti specifiche

```{r plot_tabella_ch, include = TRUE}
cas_kable(header, caption='header_CH')
```


## dati_KV_TVB

Il file contiene i dati delle prove meccaniche (Kv, trazioni) per scheda + dati LAM. 

```{r head_tvb, include = TRUE}

header <- read_delim('/data/res431/header/header_KVTVB.csv', delim=';', col_type=cols())

header <- filter(header, grab>0)
header <- header[-ncol(header)]

```

Il file è composto da `r nrow(header)` colonne, con le seguenti specifiche

```{r plot_tabella_tvb, include = TRUE}
cas_kable(header, caption='header_TVB')
```

## dati_RINV1

```{r head_rinv1, include = TRUE}

header <- read_delim('/data/res431/header/header_RINV1.csv', delim=';', col_type=cols())

header <- filter(header, grab>0)
header <- header[-ncol(header)]

```

Il file è composto da `r nrow(header)` colonne, con le seguenti specifiche

```{r plot_tabella_rinv1, include = TRUE}
cas_kable(header, caption='header_rinv1')
```

## dati_RINV2v2

```{r head_rinv2, include = TRUE}

header <- read_delim('/data/res431/header/header_RINV2v2.csv', delim=';', col_type=cols())

header <- filter(header, grab>0)
header <- header[-ncol(header)]

```

Il file è composto da `r nrow(header)` colonne, con le seguenti specifiche

```{r plot_tabella_rinv2, include = TRUE}
cas_kable(header, caption='header_rinv2')
```

*La verifica dei dati (duplicati, valori mancanti o sbagliati, ...) per ogni singola tabella è affidata al fornitore dei dati stessi*
