==> doc/05-data-discovery.Rmd <==
---
title: "Data Discovery "
subtitle: "res431"
author: "pbrunier@cogne.com"
version: 1.0 
date: "`r Sys.Date()`"
output:
  rmdformats::readthedown:
    css: ./custom.css
    df_print: paged
    gallery: no
    highlight: default
    html_document: null
    lightbox: yes
    number_sections: yes
    self_contained: yes
    thumbnails: no
editor_options:
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      warning = FALSE,
                      message = FALSE,
                      error = FALSE,
                      fig.height = 7,
                      fig.width = 10,
                      collapse = TRUE,
                      cols.print=20)
rm(list = ls(all.names = TRUE))
require(dplyr)
require(rlang)
require(kableExtra)
require(readr)
require(ggplot2)
require(reticulate)
require(rpart)
require(rpart.plot)
require(tidyr)
require(randomForest)
source('~/dev/res431/R/function.R')
```

# Analisi per odp

leggiamo i dati 
(tolgo le due colate con Ni alto)

```{python load_dati}

import pandas as pd

df = pd.read_pickle('/data/res431/res431.pkl')

df = df[df.Ni < 2].reset_index()
df = df.drop('index',axis=1)

df = df[df.T_G14 > 0].reset_index()
df = df.drop('index',axis=1)

# df = df[df.min_1050 != 'N.D.'].reset_index()
# df = df.drop('index',axis=1)
# 
# df = df[df.min_1100 != 'N.D.'].reset_index()
# df['min_1100'] = pd.to_numeric(df.min_1100)

```

<!-- # ```{r aggiungo_ID} -->
<!-- # res431 <- read_rds('/data/res431/res431.rds') -->
<!-- # # res431 <- filter(res431, Ni<2.0) -->
<!-- #  -->
<!-- # #ordino per data -->
<!-- #  -->
<!-- # res431 <- res431[order(res431$data_ins, decreasing = FALSE), ] -->
<!-- # res431 <- cbind(ID = 1:nrow(res431), res431) -->
<!-- #  -->
<!-- #          -->
<!-- # ``` -->

# Analisi dei trend

```{r trend,results='asis'}

res431 <- py$df
res431$min_1050 <- as.numeric(as.character(res431$min_1050))
res431$min_1100 <- as.numeric(as.character(res431$min_1100))

response <- 'kv'

# pl <- ggplot(res431) +
#   geom_point(aes(ID,kv)) +
#   geom_smooth(
#     aes(ID,kv),
#     color = 'magenta',
#     se = FALSE,
#     method = 'loess',
#     span=0.75, degree=2
#   ) +
#   geom_smooth(
#     aes(ID,kv),
#     color = 'blue',
#     se = FALSE,
#     method = 'lm'
#   )
# title <- paste('## Variabile: ', 'kv')
# cat(title, '\n')
# print(pl)



pl2 <- ggplot(res431) +
  geom_point(aes(data_ins,kv)) +
  geom_smooth(
    aes(data_ins,kv),
    color = 'magenta',
    se = FALSE,
    method = 'loess',
    span=0.75, degree=2
  ) +
  geom_smooth(
    aes(data_ins,kv),
    color = 'blue',
    se = FALSE,
    method = 'lm'
  )
title <- paste('## Variabile: ', 'kv')
cat(title, '\n')
print(pl2)

pl2 <- ggplot(res431) +
  geom_point(aes(colata_aod,kv)) +
  # geom_smooth(
  #   aes(colata_aod,kv),
  #   color = 'magenta',
  #   se = FALSE,
  #   method = 'loess',
  #   span=0.75, degree=2
  # ) +
  geom_smooth(
    aes(colata_aod,kv),
    color = 'blue',
    se = FALSE,
    method = 'lm'
  )
title <- paste('## Variabile: ', 'kv')
cat(title, '\n')
print(pl2)

```

# Kv contro tutti (odp)

analisi dei trend del tipo x vs y 

```{r correlazioni,results='asis'}
response <- 'kv'
vars <- res431 %>% 
  select(T_media:forno_rinv2) %>%
  names()
n <- length(vars)

for (i in seq_len(n)) {
  var_i <- vars[i]
  pl <- ggplot(res431) +
    geom_point(aes(!!sym(var_i),!!sym(response))) +
    geom_smooth(
      aes(!!sym(var_i),!!sym(response)),
      color = 'magenta',
      se = FALSE,
      method = 'loess',
      span=0.9,degree=2
    ) +
    geom_smooth(
      aes(!!sym(var_i),!!sym(response)),
      color = 'blue',
      se = FALSE,
      method = 'lm'
    )
    # geom_step(aes(index, lower_limit),color='gray',alpha=.5)
  title <- paste('## Variabile: ', var_i)
  cat(title, '\n')
  print(pl)
  cat('<p>') #necessario per terminare paragrafo
}
```

# Decision Tree (odp)

```{r tree}
dati_tree <- res431 %>% 
  select(kv:peso_batch_rinv2) 

# dati_tree <- cbind(kv,dati_tree)

fm <- rpart(formula= kv ~ . ,
            data = dati_tree,
            method = 'anova',
            cp = 0.02)

rpart.plot(fm,
           extra=101)

```

# RandomForest (odp)


```{r random_forest}

# dati_tree <- subset(dati_tree, !is.nan(dati_tree[[1]]))
dati_tree <- na.omit(dati_tree)
rf <- randomForest(formula = kv ~ . ,
                 data = dati_tree,
                 ntree = 500)

# Variable importance plot
varImpPlot(rf, main='Variable importance Plot')
# Plotting model
plot(rf,main='convergence plot')
# Importance plot
importance(rf)
```


# Analisi per triple

raggruppiamo i dati 

```{python group_dati_tripla}
dfg = df.groupby('odp').mean()
dfg['kv_std'] = df.groupby('odp').std().kv
dfg['kv_n'] = df.groupby('odp').count().kv

```

# Kv contro tutti (triple)

analisi dei trend del tipo x vs y 

```{r correlazioni tripla,results='asis'}
res431 <- py$dfg

response <- 'kv'
vars <- res431 %>% 
  select(T_media:batch_rinv2) %>%
  names()
n <- length(vars)

for (i in seq_len(n)) {
  var_i <- vars[i]
  pl <- ggplot(res431) +
    geom_point(aes(!!sym(var_i),!!sym(response))) +
    geom_smooth(
      aes(!!sym(var_i),!!sym(response)),
      color = 'magenta',
      se = FALSE,
      method = 'loess',
      span=0.75,degree=2
    ) +
    geom_smooth(
      aes(!!sym(var_i),!!sym(response)),
      color = 'blue',
      se = FALSE,
      method = 'lm'
    )
  title <- paste('### Variabile: ', var_i)
  cat(title, '\n')
  print(pl)
  cat('<p>') #necessario per terminare paragrafo
}
```

# Decision Tree (triple)

```{r tree tripla}
dati_tree <- res431 %>% 
  select(kv:peso_batch_rinv2)

fm <- rpart(formula=kv ~ .,
            data = dati_tree,
            method = 'anova',
            cp = 0.02)

rpart.plot(fm,
           extra=101)

```

# RandomForest (tripla)


```{r random_forest_triple}


# dati_tree <- subset(dati_tree, !is.nan(dati_tree[[1]]))
dati_tree <- na.omit(dati_tree)
rf <- randomForest(formula = kv ~ . ,
                 data = dati_tree,
                 ntree = 500)

# Variable importance plot
varImpPlot(rf, main='Variable importance Plot')
# Plotting model
plot(rf,main='convergence plot')
# Importance plot
importance(rf)
```



<!-- # analisi per colata -->

<!-- raggruppiamo i dati -->

<!-- ```{python group_dati_colata} -->
<!-- dfg = df.groupby('colata').mean() -->
<!-- dfg['kv_std'] = df.groupby('colata').std().kv -->
<!-- dfg['kv_n'] = df.groupby('colata').count().kv -->

<!-- ``` -->

<!-- ```{r correlazioni colate,results='asis'} -->
<!-- res431 <- py$dfg -->
<!-- response <- 'kv_valori' -->
<!-- vars <- res431 %>% -->
<!--   select(T_media:batch_rinv2) %>% -->
<!--   names() -->
<!-- n <- length(vars) -->

<!-- for (i in seq_len(n)) { -->
<!--   var_i <- vars[i] -->
<!--   pl <- ggplot(res431) + -->
<!--     geom_point(aes(!!sym(var_i),!!sym(response))) + -->
<!--     geom_smooth( -->
<!--       aes(!!sym(var_i),!!sym(response)), -->
<!--       color = 'magenta', -->
<!--       se = FALSE, -->
<!--       method = 'loess', -->
<!--       span=0.75,degree=2 -->
<!--     ) + -->
<!--     geom_smooth( -->
<!--       aes(!!sym(var_i),!!sym(response)), -->
<!--       color = 'blue', -->
<!--       se = FALSE, -->
<!--       method = 'lm' -->
<!--     ) -->
<!--   title <- paste('## Variabile: ', var_i) -->
<!--   cat(title, '\n') -->
<!--   print(pl) -->
<!--   cat('<p>') #necessario per terminare paragrafo -->
<!-- } -->
<!-- ``` -->

<!-- # Decision Tree (colate) -->

<!-- I dati sono troppo pochi per un decision tree -->

<!-- # Random Forest (colate) -->

<!-- ```{r tree colate} -->



<!-- dati_tree <- res431 %>% -->
<!--   select(kv:peso_batch_rinv2) -->

<!-- dati_tree <- na.omit(dati_tree) -->
<!-- rf <- randomForest(formula = kv ~ . , -->
<!--                  data = dati_tree, -->
<!--                  ntree = 500) -->

<!-- # Variable importance plot -->
<!-- varImpPlot(rf, main='Variable importance Plot') -->
<!-- # Plotting model -->
<!-- plot(rf,main='convergence plot') -->
<!-- # Importance plot -->
<!-- importance(rf) -->

<!-- ``` -->
